Rm pfmp amount calculator, move tests to pfmp_manager_spec
