+++ app/views/pfmps/_payment_panel.html.haml
@@ -13,5 +13,5 @@
     = number_to_currency pfmp.wage.yearly_cap
     de plafond maximum par an
   %div.fr-mb-1w
-    = number_to_currency pfmp.previously_locked_amount
+    = number_to_currency PfmpManager.new(pfmp).previously_locked_amount
     engagés par les PFMPs précédemment déclarées et validées.
